fixed some functions

+++ ft_striteri.c
@@ -6,7 +6,7 @@
 /*   By: ihalim <ihalim@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/22 10:33:16 by ihalim            #+#    #+#             */
-/*   Updated: 2024/10/27 15:35:22 by ihalim           ###   ########.fr       */
+/*   Updated: 2024/11/04 09:28:26 by ihalim           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_striteri(char *s, void (*f)(unsigned int, char *))
 {
-	size_t	i;
+	unsigned int	i;
 
 	if (!s || !f)
 		return ;
