+++ ft_strmapi.c
@@ -6,7 +6,7 @@
 /*   By: ihalim <ihalim@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/22 10:33:32 by ihalim            #+#    #+#             */
-/*   Updated: 2024/10/27 15:49:39 by ihalim           ###   ########.fr       */
+/*   Updated: 2024/11/04 09:28:05 by ihalim           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
-	char	*str;
-	size_t	i;
+	char			*str;
+	unsigned int	i;
 
 	if (!s || !f)
 		return (NULL);
